rename routes folder to screens. Moved screens.js logic to screens/index.js.
