--- src/navigation/routes/index.js
@@ -1,19 +1,0 @@
-import sidebarRoutes from './sidebar';
-import searchRoutes from './search';
-import feedRoutes from './feed';
-import libraryRoutes from './library';
-import pageRoutes from './pages';
-import authRoutes from './auth';
-import onboardingRoutes from './onboarding';
-import notificationRoutes from './notification';
-
-export {
-  sidebarRoutes,
-  searchRoutes,
-  feedRoutes,
-  libraryRoutes,
-  pageRoutes,
-  authRoutes,
-  onboardingRoutes,
-  notificationRoutes,
-};
