+++ src/navigation/screens/index.js
@@ -6,8 +6,17 @@ import App from 'kitsu/App';
 import { withNotifications } from 'kitsu/utils/notifications';
 import QuickUpdateScreen from 'kitsu/screens/QuickUpdateScreen';
 import { LightBox } from 'kitsu/screens/LightBox';
-import * as Screens from './types';
-import { feedRoutes, sidebarRoutes, searchRoutes, libraryRoutes, pageRoutes, authRoutes, onboardingRoutes, notificationRoutes } from './routes';
+
+import * as Screens from 'kitsu/navigation/types';
+
+import sidebarRoutes from './sidebar';
+import searchRoutes from './search';
+import feedRoutes from './feed';
+import libraryRoutes from './library';
+import pageRoutes from './pages';
+import authRoutes from './auth';
+import onboardingRoutes from './onboarding';
+import notificationRoutes from './notification';
 
 const routes = {
   ...feedRoutes,
@@ -23,12 +32,16 @@ const routes = {
   [Screens.INITIAL]: App,
 };
 
+
+function registerComponent(name, callback) {
+  Navigation.registerComponentWithRedux(name, callback, Provider, store);
+}
+
+/**
+ * Register the screens for React Native Navigation
+ */
 export function registerScreens() {
   Object.keys(routes).forEach((key) => {
     registerComponent(key, () => withNotifications(routes[key]));
   });
-}
-
-function registerComponent(name, callback) {
-  Navigation.registerComponentWithRedux(name, callback, Provider, store);
 }
